server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy>=2.0.25
pymysql==1.1.0
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pydantic>=2.5.3
python-multipart==0.0.6
python-dotenv==1.0.1
